R48xx: allow sending 32 bit values
all values reported by the unit are 32 bit values, so it should be very
safe to assume that values being sent to the unit should also be 32 bit.

+++ include/gridcharger/huawei/HardwareInterface.h
@@ -61,7 +61,7 @@ private:
 
     std::unique_ptr<DataPointContainer> _upData = nullptr;
 
-    std::queue<std::pair<HardwareInterface::Setting, uint16_t>> _sendQueue;
+    std::queue<std::pair<HardwareInterface::Setting, uint32_t>> _sendQueue;
 
     static unsigned constexpr _maxCurrentMultiplier = 20;
